Refactor Sidebar, add WelcomeModal, and update UI theme
- Refactored `src/components/Sidebar.tsx` into modular sub-components in `src/components/sidebar/`.
- Implemented `WelcomeModal` to show on first access.
- Updated `src/index.css` with brand colors matching the logo (Blue/Gold).
- Updated mobile header styling in `src/App.tsx`.
- Updated `README.md` with tech stack and deployment info.
